+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -29,6 +29,8 @@ import com.mars_sim.core.building.task.MaintainBuildingMeta;
 import com.mars_sim.core.maintenance.MaintainRobotMeta;
 import com.mars_sim.core.malfunction.task.RepairMalfunctionMeta;
 import com.mars_sim.core.manufacture.task.ManufacturingMetaTask;
+import com.mars_sim.core.mission.task.PlanMissionMeta;
+import com.mars_sim.core.mission.task.ReviewMissionPlanMeta;
 import com.mars_sim.core.person.ai.task.meta.AnalyzeMapDataMeta;
 import com.mars_sim.core.person.ai.task.meta.ConnectOnlineMeta;
 import com.mars_sim.core.person.ai.task.meta.ConsolidateContainersMeta;
@@ -38,14 +40,12 @@ import com.mars_sim.core.person.ai.task.meta.DoInventoryMeta;
 import com.mars_sim.core.person.ai.task.meta.EatDrinkMeta;
 import com.mars_sim.core.person.ai.task.meta.ListenToMusicMeta;
 import com.mars_sim.core.person.ai.task.meta.MeetTogetherMeta;
-import com.mars_sim.core.person.ai.task.meta.PlanMissionMeta;
 import com.mars_sim.core.person.ai.task.meta.PlayHoloGameMeta;
 import com.mars_sim.core.person.ai.task.meta.ReadMeta;
 import com.mars_sim.core.person.ai.task.meta.RecordActivityMeta;
 import com.mars_sim.core.person.ai.task.meta.RelaxMeta;
 import com.mars_sim.core.person.ai.task.meta.ReportMissionControlMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewJobReassignmentMeta;
-import com.mars_sim.core.person.ai.task.meta.ReviewMissionPlanMeta;
 import com.mars_sim.core.person.ai.task.meta.SleepMeta;
 import com.mars_sim.core.person.ai.task.meta.TeachMeta;
 import com.mars_sim.core.person.ai.task.meta.WorkoutMeta;
